战报优化 - 战报优化提交

+++ mw_service/src/main/java/com/gryphpoem/game/zw/resource/domain/p/DbCrossMap.java
@@ -81,7 +81,7 @@ public class DbCrossMap implements DbSerializeId{
     }
 
     @Override
-    public int getSerializeIdId() {
+    public long getSerializeIdId() {
         return mapId;
     }
